+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - `WebBrowserProvider` can display custom messages
-- Support deleting Task
+- Support deleting Task and TaskList
 
 ### Changed